requirements.txt
fastapi==0.109.0
fastapi-health==0.4.0
uvicorn==0.23.2
jinja2==3.1.3
python-multipart==0.0.9
passlib==1.7.4
mysql-connector-python==8.3.0

python-dotenv==1.0.0